setup.sh

+++ bin/setup.sh
@@ -1,10 +1,16 @@
 #!/bin/bash
 
-rm bin/setup.sh
-rm `find -name '.removeme'`
-mv tpl.gitignore .gitignore
-mv tpl.README.md README.md
-rm LICENSE
-mkdir var/media
-mkdir var/static
-pip install -r requirements/dev.txt
+SELF="$( cd "$( dirname "$0" )" && pwd )/bin/setup.sh"
+
+if [ -f $SELF ]; then
+	rm `find -name '.removeme'`
+	mv tpl.gitignore .gitignore
+	mv tpl.README.md README.md
+	rm LICENSE
+	mkdir var/media
+	mkdir var/static
+	pip install -r requirements/dev.txt
+	rm bin/setup.sh
+else
+	echo "You should run setup from project root. Aborting."
+fi
